Delivering for the tester and the customer.
In order to be able to deliver to the customer and the tester the Info.plist has to be modified. The application, depending whether it
a tester a pplication or a developer on, shall look in a different place in the server for its update. So, the infor.Plist shall be
adapted before compile.

Add the marketing string version to the generate appcast xml file.